Fix HWB conversion
Not only were we still using a 1 when we should have used 100,
but when we moved to this algorithm, we forgot to keep the white and
black correction if the sum exceeded 100.

Just normalize to [0..1] and then correct if the white and black sum
exceeds 100. Finally, use the same algorithm as before, just use the
proper values.