Don't force hue to zero in translation of lab to lch
The workaround in colorjs.io that forces hue to zero during conversion
of lab to lch actually can throw off some conversions to a small degree.
They do this by setting the hue to NaN, which is treated as zero in
color space conversion. I believe the intention of this is mainly for
color interpolation. We calculate whether we should consider the hue
during interpolation at time of interpolation, so we don't need this
logic.