More gamut fixes/adjustments
By default use no threshold and force 0 threshold in all fit methods.

Also, disable using other spaces to determine if in_gamut. I'm not
convinced it is needed if we are checking both anyways. And my
assumption that fitting must be done in the same space as the check may
be wrong as well.

We've already fixed the issues of conversions and gamut checks by using
a precise gamut check, so really the only issue is whether spaces like
HSL should be fit in sRGB or in there own space. I'm thinking it may be
the latter.