Force subclassed objects to be converted to current class and cleanup
A subclassed object may or may not have modified other functional
behavior. So if two classes are are not instances of the same class,
the logic could be quite different. For instance, if a subclassed object
modifies luminance, and the default color class is used with this
modified class, the contrast ratio may not make sense. So normalizing
the class objects to whatever if the calling class is will prevent
related issues.

In most cases, I imagine consistent classes will be used, but if not, at
least results will be ensured to be sane.

Additionally some simplification of logic is some areas was made