Fix some slightly off calculations with din99o
White would never translate to 100% lightness in din99o.  Turns out this
is because of a rounding issue. Solving backwards to see how we get the
303.67 constant, we can see that the natural log is taken of 1.39 when
plugging in for a full, white lightness. The actual result of of the
calculation leaves us with 303.67 and some change. This more precise
value fixes the white translation.