Fix inconsistencies with conversion
We use our own rgb/xyz values as recommended in OkLab, but in okhsl and
okhsv, we were using their matrix for srgb linear. We should have
matched ours. This caused a slight inconsistency. This is now fixed
and the round trip test passes again, without any kind of patch.