Recalculate XYZ Oklab matrix using newer sRGB matrix
The Oklab document seems to give the impression that both the XYZ matrix
and the sRGB Linear matrix were both updated, but this is not actually
true, only the latter was updated. You must recalculate taking the dot
of the "XYZ to sRGB" matrix you are using and the "sRGB to LMS" matrix
provided by the Oklab article.