Update bounds object
Before we had bounding objects wrapped around typed floats, but the
reality is that we don't need specifically typed floats as all values
with a boundary object should be of the same type: angle vs percent,
etc. As we are now used typed tuples, just add a flag member where we
can set angle flags, percent flags, etc. Now we can just check the flag
to see what type of content and easily expand it if necessary.

Rename RANGE to BOUNDS.