Clamp lower end of chroma to 0 at the channel level
It is impossible, mathematically to get something less than zero, no
conversion or anything will yield such values.

Less than zero is only achieved via user input, and if we try to display
it, we have to normalize it to zero anyways. If we try to convert from
it, we also have to clamp it. So there is no harm in auto clamping it
anytime a user attempts a value less than zero.