Fix matrix division issue and remove all recursive calls in mult/div
I thought we had tested all cases and compared them to numpy, but we
apparently missed a particular case and divide was mixing up numerator
and denominator.

All cases have formally been compared to numpy and should operate as
expected.

The particular case is not used anywhere in our code, so there is no
need to release a bugfix release at this time, but the fix will be
included in future release in case we ever do utilize the utility in
such a manner as to exploit this case.