Slight refactor to base interpolate plugin in regards to extrapolation
We do not currently do extrapolation in any of our interpolators, but
we shouldn't make it difficult to do so.

The desired point on the interpolation line should remain unmodified
through the call chain up until the a user would write custom
interpolation code, e.g. the `interpolate` function of the
`Interpolator` class.

Once there, they can decide if they'd like to not clamp the value. This
can be useful if a periodic, closed loop interpolator was desired. In
this case, it may be beneficial to allow values beyond 0 and 1, and wrap
them.