Slight refactor to base interpolate plugin in regards to extrapolation (#226)
* Slight refactor to base interpolate plugin in regards to extrapolation

We do not currently do extrapolation in any of our interpolators, but
we shouldn't make it difficult to do so.

The desired point on the interpolation line should remain unmodified
through the call chain up until the a user would write custom
interpolation code, e.g. the `interpolate` function of the
`Interpolator` class.

Once there, they can decide if they'd like to not clamp the value. This
can be useful if a periodic, closed loop interpolator was desired. In
this case, it may be beneficial to allow values beyond 0 and 1, and wrap
them.

* Add an extrapolate option in interpolation functions

- Add extrapolate option
- Adjust how spline methods extrapolate values
- Consolidate easing logic into to one place
- Sight refactor to better improve subclassing of interpolate plugin

* Mypy fixes

* Add extrapolation tests and add small fix-up to B-spline alpha

Floating point math can sometimes make an interpolation of multiple,
fully opaque colors return very, very slight transparent colors.