Some color spaces need specific hues for better achromatic conversion
Because of the complexity of some of the lab-like color space
algorithms, some spaces do not often land precisely with exact zeros for
a and b, which means their polar forms do not have perfect zero for
chroma. This means if you set chroma to zero and convert back, you will
often not get exactly the same value. These algorithms will generally
have a specific hue that gets the best result. It will vary slightly,
but will round out to about the same value. This provides a little
better conversions back to achromatic values. We will use the hue
at white before we normalize it to NaN. Generally, for the spaces that
don't hover at zero, this improves achromatic conversions.