More achromatic adjustments
Don't manipulate the algorithm to give better values for conversions
as we introduce inaccuracies when doing so. If conversion between
spaces yields slightly out of bounds values, it is up to the user to
gamut map them back in.

Abstract the dynamic achromatic logic and apply it to JzCzhz

Pre-calculate achromatic responses for HCT, JMh, and JzCzhz.