Even better achromatic handling
Undefined chroma will also use a neutral value for a given lightness.
Lab colors will use the best neutral value for a given lightness when
they are undefined also. Achromatic round trip values are extremely
better. We were able to up the worst offender HCT to from 3 to the
default precision of 5 for acceptable round trips.

Oklab and OKLCh also now always use the best value for chroma, hue, a,
or b. Serialization has a new `css_undefined` option to force a
serialization that matches CSS if it is strongly desired.