Even faster HCT conversion
We curve fitted CAM16 J vs HCT T to come up with a polynomial to guess
the best J possible for Newton Raphson. If the color has a very wide
gamut, we will just correct the luminance in XYZ instead as the
J vs T relationship changes as chroma gets larger causing the initial J
guess to get worse in some regions as chroma increase.