Small fix for HWB 3D Plotly diagrams
The center point on the bottom disc of the HWB cylinder is a point that
can be any color depending on the hue. When modeling this color, one
variant of this point is picked as the center point, usually red. When
we average the faces close to this point, we get a small circle where
all the colors have a red bias, due to the center point being red.
Avoid this by interpolating very close to the center point without
explicitly interpolating the point at the center of the disc. The points
will still be connected at the center, but now we will make that point
so small, that it won't be noticeable without zooming in at extreme
levels. This is not a problem for matplotlib surface diagrams as their
mesh is calculated differently, and those surface colors are calculated
differently. If we migrate matplotlib to a "trimesh" diagram opposed to
the more simple square mesh of the surface plot, we'd have to do the
same workaround.