Fix inv row sorting once and for all
We should now be able to sort the rows such that the diagonal never
have zeros in them unless the matrix is invertible. We seem to be
able to handle 30 x 30 2D matrices, much larger than we would ever
need. There are certainly cases that could stress this, but from the
perspective of what we need for colors, this should be good enough.