Deprecate is_nan, no_nan, and no_nans
We don't handle nans with the naive no_nan and no_nans anymore.
It is faster to use `isnan` directly, and if we want to check an array
we've provided a new alg.isnan method. It may still be faster to
use a specialized nan check, but for convenience, alg.isnan is good.