Use math constants directly from `math`
There is no reason to explicitly provide our own in algebra as nan, inf,
and tau are all in math. We will leave NaN and INF for now as they've
been there a good while, but tau wasn't really used anywhere until
recently. We also never used EPSILON.

We obviously will still provide coloraide.NaN as that is a convenience
constant and part of the API.