Use CSS algorithm for HWB instead of HSV approach
HSV approach is correct, but CSS employs an approach that uses HSL
and sRGB that is reversible in the negative lightness direction all
while resolving the RGB values to the exact same values.

While negative lightness colors are not useful, we always prefer the
approach with better overall round tripping for predictable conversions.