Make internal logic for CAT less confusing
It doesn't really matter the order of division as long as you get the
source and destination context correct and use the proper matrix:
forward vs inverse. But code wise, this was not clear.

While we sort the source and destination white point and cache the
forward and inverse matrices and then determine which to use based on
specified order, at a glance it can be a bit confusing. Explicitly
state the source and destination.