Apply proper negative lightness handling in DIN99o
We really shouldn't try to be clever here by trying to mirroring the
behavior for negative lightness. Instead, just let the algorithm as
specified handle it, and when it it breaks down (somewhere past -200)
we can just return NaN for now. In most cases, this will be more than
sufficient. Anyone hitting this limit is pushing the space well pass
reasonable limits.