Okhsl and Okhsv are gamut mapped in their own gamut
Okhsl and Okhsv do not represent a proper sRGB gamut. It is close, but
not exact. Its bounds clip a little of sRGB but also include a little
past sRGB. It doesn't make sense to gamut map it directly to sRGB.
Users should expressly gamut in sRGB if that is desired.