Break vectorize2 up into vectorize1 and vectorize2
These are tailored to to specifically handle functions with 1
vectorized input or two respectively. Before, vectorize2 could handle
both, but you could pass two inputs to a function that only takes one.

For faster, no frills vectorization, these can be used and can be made
faster if you pass in the dimensions to allow it to skip analyzing
shape. vectorize is still available for multi inputs of any length, but
is slower.