Leave ray trace defined as experimental for now.
No need to rush status. It is quite stable to use with the caveat that
it has some weaknesses with a few spaces in adaptive mode, but this is
to be expected due to how it works. It may be that we can tighten that
up with some changes in the future, but we'll leave it as experimental
for now.