create is very confusing as a variable name, use color_cls
`create` was used as we pass the color class so we can create colors
with the current color object. This was used in non class functions
or plugins removed from having access to the color object, but this
is a confusing name. We should call it what it is. This is mostly a
benign, internal change, but it does affect the interpolation plugin as
we must rename an input parameter making this a breaking change.