Improve the ray trace algorithm to yield better results in all spaces
The ray trace approach was designed focusing mainly on Oklab/OkLCh.
Further evaluation shows that some spaces twist a little different
through the RGB spaces and can yield results not as good, CIELCh
being an example.

Adjust the algorithm so that if a point is found along the reduction
path below the gamut surface, that point becomes the new anchor.

This generally tightness the range in a few passes and prevent some
spaces from having larger than expected hue deviations.