Optimized versions of basic math functions for length 3 arrays
Majority of array operations in ColorAide are performed on 3x3 matrices
and length 3 vectors. Knowing this, we can provide math functions that
are less generalized and avoid loops creating more performant
operations.

Combine the functionality of vectorize1 and vectorize2 such that we now
just have vectorize2 with the parameter `params` which can be set to 1
or 2 (2 being the default) and will and can vectorize functions that
accept 1 or 2 array parameters. vectorize1 is now deprecated. These
functions, as always, are still limited to dimensions of 2 or less.

Additionally, provide a new `only_x3` parameter to vectorize2 that will
allow us to use an even more optimized logic when vectorizing a function
with the limitation that the matrix or vector must have dimension
lengths of 3.

Lastly, create new optimized functions: matmul_x3, dot_x3, multiply_x3,
divide_x3, add_x3, and subtract_x3 which require arrays of dimensions
less than 2 and dimension lengths of 3 and will perform much faster.

Replace all applicable calls with these new x3 variants of basic math
functions.