Add a way to install modules in notebooks
Because we are parsing Python content to deliver rich color previews,
we can't simply use micropip and await. When the code is processed,
we are already in a loop and we have to run, at least when parsing
Markdown, in synchronous functions.

So, if processing code in a live code block, process
`# micropip-require: module1 module2` as the first line.

If processing a live notebook, use `<!-- micropip-require: module -->`,
but at the top of the Markdown page and all code blocks will see it.
You can also include it in the code blocks individually as it should be
cached, that way if you want to share a block, it would have everything
needed.