Improve RYB biased handling
A more accurate cubic solving approach is not desired in RYB.
Using a more accurate cubic solver on less accurate inverse trilinear
interpolations can push some solutions to be more inaccurate. This was
observed near RYB [1, 1, 0]. It is better that we match the cubic
solver accuracy to the inverse trilinear interpolation accuracy which
is about 1e-6.