Fix XYB conversion (#463)
The default form of XYB has X = B = 0 for achromatic colors. This was
done by altering the XYB transform matrix to do the LMS to XYB
conversion and apply B - Y, but we had done so incorrectly. The matrix
that incorporates this change should have been:

XYB_LMS_TO_XYB = [
    [0.5, -0.5, 0.0],
    [0.5, 0.5, 0.0],
    [-0.5, -0.5, 1.0]
]

We mistakenly applied the logic in reverse applying to the forward
transform which should have been in the inverse transform and vice
versa. It resulted in the same X = B = 0, but non achromatic colors
were not quite right.

With all of that said, instead of using a clever matrix to do the XYB
transform and subtract Y from B, we now use the standard matrix and then
just subtract Y from B at the end. We verified both were equivalent, but
the later approach makes it more clear what we are doing.