Docs: Fix issue with 3D plots demo
Plotly seems to sometimes not show the bottom of HSL and some other
RGB cylinders at certain, specific resolutions. It doesn't seem like it
is specifically anything we are doing, but a quirk. It seems if we
use a very small value close to 0 for lightness (1e-16), the issue goes
away.