Speed up conversion by avoiding unnecessary chromatic adaptation (#478)
When converting, we currently apply chromatic adaptation when converting
to and from XYZ D65, our base space through which all chromatic adaption
occurs. There are many times when we are converting to and from XYZ D65
from a space that is already in XYZ D65.

When we calculate our conversion chain, and are converting to or from
XYZ D65, check if the white points are the same. If so, skip chromatic
adaptation.