Ignore ACEScct and ACEScc as demo-able gamuts in 3D model demo
We had considered separating RGB from Regular spaces and using that to
filter out abnormal RGB models with strange ranges, but decided against
it and re-defined Regular to mean prism spaces, which RGB space
definitely are. So add ACEScc and ACEScct back to the ignore list which
is the easiest way to avoid plotting them.