Use the appropriate transfer function for Rec. 2020 (#488)
* Use the appropriate transfer function for Rec. 2020

CSS has resolved to change Rec. 2020 to use the OETF in BT.1886 with a
gamma of 2.4.

This changes Rec. 2020 to comply and adds a new space called
`rec2020-oetf` that continues to the use the old one as specified in
BT.2020.

* Simplify transfer function

When using the CSS defaults, the actual transfer function becomes a
simple power or inverse power of 2.4 (depending on direction of
translation).

* Note which Rec. 2020 space is scene-referred vs display-referred

* Update changelog

* Fix changelog wording

* Update version